Updating multinet build handler. Add clean handler for multinet

+++ emulators/multinet/build.sh
@@ -13,8 +13,18 @@ echo $SCRIPT_DIR
 MULTINET_LOCATION="https://github.com/intracom-telecom-sdn/multinet.git"
 
 if [ ! -d $SCRIPT_DIR"/multinet" ]; then
-	git clone $MULTINET_LOCATION $SCRIPT_DIR"/multinet"
-	rm -rf $SCRIPT_DIR"/multinet/.git"
-	mv $SCRIPT_DIR/multinet/* $SCRIPT_DIR
-	rm -rf $SCRIPT_DIR/multinet
+    git clone $MULTINET_LOCATION $SCRIPT_DIR"/multinet"
+    if [ $? -ne 0 ]; then
+        echo "[build.sh] Cloning multinet failed. Exiting ..."
+        exit 1
+    fi
+    rm -rf $SCRIPT_DIR"/multinet/.git"
+    mv $SCRIPT_DIR/multinet/* $SCRIPT_DIR
+    if [ $? -ne 0 ]; then
+        echo "[build.sh] Moving multinet files failed. Exiting ..."
+        exit 1
+    fi
+    rm -rf $SCRIPT_DIR/multinet
 fi
+
+echo "[build.sh] Building multinet completed successfully"
